APIGOV-22856 - code gen, update float32 to float64

+++ pkg/apic/apiserver/models/catalog/v1/model_quota_spec_limit_type_loose.go
@@ -13,6 +13,7 @@ package v1
 type QuotaSpecLimitTypeLoose struct {
 	Type string `json:"type"`
 	// The limit of the unit that is provided.
-	Value    float32                         `json:"value"`
+	// GENERATE: The following code has been modified after code generation
+	Value    float64                         `json:"value"`
 	Overages QuotaSpecLimitTypeLooseOverages `json:"overages"`
 }
